# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Ini adalah dokumentasi API untuk backend Pamagi.",
        "title": "Pamagi API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/words": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Menampilkan daftar kosakata dengan fitur filter kombo (limit 20)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Words"
                ],
                "summary": "Ambil Daftar Kosakata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by Category ID",
                        "name": "category_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by Part of Speech (e.g., NOUN)",
                        "name": "part_of_speech",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by Favorite status",
                        "name": "is_favorite",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Filter by Bookmark status",
                        "name": "is_bookmarked",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by Start Date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by End Date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sorting (newest, oldest, a_z, z_a)",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Nomor Halaman (Default: 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Jumlah Data per Halaman (Default: 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.WordResponse"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Menyimpan kosakata beserta contoh kalimat opsional",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Words"
                ],
                "summary": "Tambah Kosakata Baru",
                "parameters": [
                    {
                        "description": "Data Kosakata",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateWordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/words/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Mengambil satu data kosakata beserta relasi kategori dan contoh kalimatnya",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Words"
                ],
                "summary": "Ambil Detail Kosakata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kosakata",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.WordResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Menghapus satu kosakata beserta relasi kategori dan contoh kalimatnya secara permanen",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Words"
                ],
                "summary": "Hapus Kosakata",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kosakata",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/domain.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/words/{id}/bookmark": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Menandai atau menghapus tanda bookmark pada suatu kata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Words"
                ],
                "summary": "Ubah Status Bookmark",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kosakata",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/words/{id}/favorite": {
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Menandai atau menghapus tanda favorit pada suatu kata",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Words"
                ],
                "summary": "Ubah Status Favorit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID Kosakata",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/domain.SuccessResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "domain.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "dto.CategoryRes": {
            "type": "object",
            "properties": {
                "icon": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "dto.CreateWordRequest": {
            "type": "object",
            "required": [
                "part_of_speech",
                "russian_word",
                "translation"
            ],
            "properties": {
                "category_ids": {
                    "description": "Array ID kategori (Opsional)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "examples": {
                    "description": "Array contoh kalimat (Opsional)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ExampleRequest"
                    }
                },
                "part_of_speech": {
                    "type": "string"
                },
                "russian_word": {
                    "type": "string"
                },
                "translation": {
                    "type": "string"
                }
            }
        },
        "dto.ExampleRequest": {
            "type": "object",
            "required": [
                "russian_sentence",
                "translated_sentence"
            ],
            "properties": {
                "russian_sentence": {
                    "type": "string"
                },
                "translated_sentence": {
                    "type": "string"
                }
            }
        },
...
        "dto.WordExampleRes": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "russian_sentence": {
                    "type": "string"
                },
                "translated_sentence": {
                    "type": "string"
                }
            }
        },
        "dto.WordResponse": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.CategoryRes"
                    }
                },
                "created_at": {
                    "type": "string"
                },
                "examples": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.WordExampleRes"
                    }
                },
                "id": {
                    "type": "string"
                },
                "is_bookmarked": {
                    "type": "boolean"
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "part_of_speech": {
                    "type": "string"
                },
                "russian_word": {
                    "type": "string"
                },
                "translation": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}